feat: Remove install script
